refactoring: create regional_backend_service class
