+++ vm_network_migration/module_helpers/backend_service_helper.py
@@ -15,10 +15,10 @@
 """
 from googleapiclient.http import HttpError
 from vm_network_migration.errors import *
-from vm_network_migration.modules.backend_service_modules.external_backend_service import ExternalBackendService
-from vm_network_migration.modules.backend_service_modules.internal_backend_service import InternalBackendService
+from vm_network_migration.modules.backend_service_modules.external_global_backend_service import ExternalBackendService
+from vm_network_migration.modules.backend_service_modules.internal_regional_backend_service import InternalBackendService
 from vm_network_migration.utils import initializer
-from vm_network_migration.modules.backend_service_modules.internal_self_managed_backend_service import InternalSelfManagedBackendService
+from vm_network_migration.modules.backend_service_modules.internal_self_managed_global_backend_service import InternalSelfManagedBackendService
 
 
 class BackendServiceHelper:
